erts: Move sys drivers to another file
